tests/requirements.txt
mock==2.0.0
djangorestframework==3.14
